Rename JSErrorInfo to FormatStackTraceForJS
The new name better describes what the code does - it formats stack traces
for JavaScript error objects.
