+++ src/bun.js/bindings/FormatStackTraceForJS.h
